+++ src/main/java/client/exception/EosApiError.java
@@ -1,9 +1,7 @@
 package client.exception;
 
-
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
-
 
 public class EosApiError {
 
@@ -39,17 +37,15 @@ public class EosApiError {
 
     @Override
     public String toString() {
-        return new ToStringBuilder(this, ToStringStyle.SHORT_PREFIX_STYLE)
-                .append("message", message)
-                .append("code", code)
-                .toString();
+        return new ToStringBuilder(this, ToStringStyle.SHORT_PREFIX_STYLE).append("message", message)
+                .append("code", code).toString();
     }
 
     public String getDetailedMessage() {
         return error == null ? message : message + ": " + error.getWhat();
     }
 
-    public Integer getEosErrorCode() {
+    public Long getEosErrorCode() {
         return error == null ? null : error.getCode();
     }
 }
