+++ src/main/java/client/impl/EosApiRestClientImpl.java
@@ -1,6 +1,5 @@
 package client.impl;
 
-
 import client.EosApiRestClient;
 import client.domain.common.WalletKeyType;
 import client.domain.common.transaction.SignedPackedTransaction;
@@ -20,7 +19,6 @@ import client.domain.response.history.controlledaccounts.ControlledAccounts;
 import client.domain.response.history.keyaccounts.KeyAccounts;
 import client.domain.response.history.transaction.Transaction;
 
-
 import java.util.*;
 
 public class EosApiRestClientImpl implements EosApiRestClient {
@@ -31,45 +29,49 @@ public class EosApiRestClientImpl implements EosApiRestClient {
 
     private final EosHistoryApiService eosHistoryApiService;
 
-    public EosApiRestClientImpl(String baseUrl){
+    public EosApiRestClientImpl(String baseUrl) {
         eosWalletApiService = EosApiServiceGenerator.createService(EosWalletApiService.class, baseUrl);
         eosChainApiService = EosApiServiceGenerator.createService(EosChainApiService.class, baseUrl);
         eosHistoryApiService = EosApiServiceGenerator.createService(EosHistoryApiService.class, baseUrl);
     }
 
-    public EosApiRestClientImpl(String walletBaseUrl, String chainBaseUrl, String historyBaseUrl){
+    public EosApiRestClientImpl(String walletBaseUrl, String chainBaseUrl, String historyBaseUrl) {
         eosWalletApiService = EosApiServiceGenerator.createService(EosWalletApiService.class, walletBaseUrl);
         eosChainApiService = EosApiServiceGenerator.createService(EosChainApiService.class, chainBaseUrl);
         eosHistoryApiService = EosApiServiceGenerator.createService(EosHistoryApiService.class, historyBaseUrl);
     }
 
     @Override
-    public ChainInfo getChainInfo(){
+    public ChainInfo getChainInfo() {
         return EosApiServiceGenerator.executeSync(eosChainApiService.getChainInfo());
     }
 
     @Override
-    public Block getBlock(String blockNumberOrId){
-        return EosApiServiceGenerator.executeSync(eosChainApiService.getBlock(Collections.singletonMap("block_num_or_id", blockNumberOrId)));
+    public Block getBlock(String blockNumberOrId) {
+        return EosApiServiceGenerator
+                .executeSync(eosChainApiService.getBlock(Collections.singletonMap("block_num_or_id", blockNumberOrId)));
     }
 
     @Override
-    public Account getAccount(String accountName){
-        return EosApiServiceGenerator.executeSync(eosChainApiService.getAccount(Collections.singletonMap("account_name", accountName)));
+    public Account getAccount(String accountName) {
+        return EosApiServiceGenerator
+                .executeSync(eosChainApiService.getAccount(Collections.singletonMap("account_name", accountName)));
     }
 
     @Override
-    public Abi getAbi(String accountName){
-        return EosApiServiceGenerator.executeSync(eosChainApiService.getAbi(Collections.singletonMap("account_name", accountName)));
+    public Abi getAbi(String accountName) {
+        return EosApiServiceGenerator
+                .executeSync(eosChainApiService.getAbi(Collections.singletonMap("account_name", accountName)));
     }
 
     @Override
-    public Code getCode(String accountName){
-        return EosApiServiceGenerator.executeSync(eosChainApiService.getCode(Collections.singletonMap("account_name", accountName)));
+    public Code getCode(String accountName) {
+        return EosApiServiceGenerator
+                .executeSync(eosChainApiService.getCode(Collections.singletonMap("account_name", accountName)));
     }
 
     @Override
-    public TableRow getTableRows(String scope, String code, String table){
+    public TableRow getTableRows(String scope, String code, String table) {
         LinkedHashMap<String, String> requestParameters = new LinkedHashMap<>(7);
 
         requestParameters.put("scope", scope);
@@ -81,7 +83,7 @@ public class EosApiRestClientImpl implements EosApiRestClient {
     }
 
     @Override
-    public List<String> getCurrencyBalance(String code, String accountName, String symbol){
+    public List<String> getCurrencyBalance(String code, String accountName, String symbol) {
         LinkedHashMap<String, String> requestParameters = new LinkedHashMap<>(3);
 
         requestParameters.put("code", code);
@@ -92,7 +94,7 @@ public class EosApiRestClientImpl implements EosApiRestClient {
     }
 
     @Override
-    public AbiBinToJson abiBinToJson(String code, String action, String binargs){
+    public AbiBinToJson abiBinToJson(String code, String action, String binargs) {
         LinkedHashMap<String, String> requestParameters = new LinkedHashMap<>(3);
 
         requestParameters.put("code", code);
@@ -105,26 +107,29 @@ public class EosApiRestClientImpl implements EosApiRestClient {
     @Override
     @SuppressWarnings("unchecked")
     public <T> AbiJsonToBin abiJsonToBin(String code, String action, T args) {
-        return EosApiServiceGenerator.executeSync(eosChainApiService.abiJsonToBin(new AbiJsonToBinRequest(code, action, args)));
+        return EosApiServiceGenerator
+                .executeSync(eosChainApiService.abiJsonToBin(new AbiJsonToBinRequest(code, action, args)));
     }
 
     @Override
-    public PushedTransaction pushTransaction(String compression, SignedPackedTransaction packedTransaction){
-        return EosApiServiceGenerator.executeSync(eosChainApiService.pushTransaction(new PushTransactionRequest(compression, packedTransaction, packedTransaction.getSignatures())));
+    public PushedTransaction pushTransaction(String compression, SignedPackedTransaction packedTransaction) {
+        return EosApiServiceGenerator.executeSync(eosChainApiService.pushTransaction(
+                new PushTransactionRequest(compression, packedTransaction, packedTransaction.getSignatures())));
     }
 
     @Override
-    public List<PushedTransaction> pushTransactions(List<PushTransactionRequest> pushTransactionRequests){
+    public List<PushedTransaction> pushTransactions(List<PushTransactionRequest> pushTransactionRequests) {
         return EosApiServiceGenerator.executeSync(eosChainApiService.pushTransactions(pushTransactionRequests));
     }
 
     @Override
-    public RequiredKeys getRequiredKeys(PackedTransaction transaction, List<String> keys){
-        return EosApiServiceGenerator.executeSync(eosChainApiService.getRequiredKeys(new RequiredKeysRequest(transaction, keys)));
+    public RequiredKeys getRequiredKeys(PackedTransaction transaction, List<String> keys) {
+        return EosApiServiceGenerator
+                .executeSync(eosChainApiService.getRequiredKeys(new RequiredKeysRequest(transaction, keys)));
     }
 
     @Override
-    public Map<String, CurrencyStats> getCurrencyStats(String code, String symbol){
+    public Map<String, CurrencyStats> getCurrencyStats(String code, String symbol) {
         LinkedHashMap<String, String> requestParameters = new LinkedHashMap<>(2);
 
         requestParameters.put("code", code);
@@ -134,27 +139,27 @@ public class EosApiRestClientImpl implements EosApiRestClient {
     }
 
     @Override
-    public String createWallet(String walletName){
+    public String createWallet(String walletName) {
         return EosApiServiceGenerator.executeSync(eosWalletApiService.createWallet(walletName));
     }
 
     @Override
-    public void openWallet(String walletName){
+    public void openWallet(String walletName) {
         EosApiServiceGenerator.executeSync(eosWalletApiService.openWallet(walletName));
     }
 
     @Override
-    public void lockWallet(String walletName){
+    public void lockWallet(String walletName) {
         EosApiServiceGenerator.executeSync(eosWalletApiService.lockWallet(walletName));
     }
 
     @Override
-    public void lockAllWallets(){
+    public void lockAllWallets() {
         EosApiServiceGenerator.executeSync(eosWalletApiService.lockAll());
     }
 
     @Override
-    public void unlockWallet(String walletName, String walletPassword){
+    public void unlockWallet(String walletName, String walletPassword) {
         List<String> requestFields = new ArrayList<>(2);
 
         requestFields.add(walletName);
@@ -163,7 +168,7 @@ public class EosApiRestClientImpl implements EosApiRestClient {
     }
 
     @Override
-    public void importKeyIntoWallet(String walletName, String key){
+    public void importKeyIntoWallet(String walletName, String key) {
         List<String> requestFields = new ArrayList<>(2);
 
         requestFields.add(walletName);
@@ -172,43 +177,46 @@ public class EosApiRestClientImpl implements EosApiRestClient {
     }
 
     @Override
-    public List<String> listWallets(){
+    public List<String> listWallets() {
         return EosApiServiceGenerator.executeSync(eosWalletApiService.listWallets());
     }
 
     @Override
-    public List<List<String>> listKeys(String walletName, String password){
+    public List<List<String>> listKeys(String walletName, String password) {
         List<String> requestFields = Arrays.asList(walletName, password);
-       return EosApiServiceGenerator.executeSync(eosWalletApiService.listKeys(requestFields));
+        return EosApiServiceGenerator.executeSync(eosWalletApiService.listKeys(requestFields));
     }
 
     @Override
-    public List<String> getPublicKeys(){
+    public List<String> getPublicKeys() {
         return EosApiServiceGenerator.executeSync(eosWalletApiService.getPublicKeys());
     }
 
     @Override
-    public SignedPackedTransaction signTransaction(PackedTransaction packedTransaction, List<String> publicKeys, String chainId) {
-        return EosApiServiceGenerator.executeSync(eosWalletApiService.signTransaction(new SignTransactionRequest(packedTransaction, publicKeys, chainId)));
+    public SignedPackedTransaction signTransaction(PackedTransaction packedTransaction, List<String> publicKeys,
+            String chainId) {
+        return EosApiServiceGenerator.executeSync(eosWalletApiService
+                .signTransaction(new SignTransactionRequest(packedTransaction, publicKeys, chainId)));
     }
 
     @Override
-    public void setWalletTimeout(Integer timeout){
+    public void setWalletTimeout(Long timeout) {
         EosApiServiceGenerator.executeSync(eosWalletApiService.setTimeout(timeout));
     }
 
     @Override
-    public String signDigest(String digest, String publicKey){
+    public String signDigest(String digest, String publicKey) {
         return EosApiServiceGenerator.executeSync(eosWalletApiService.signDigest(Arrays.asList(digest, publicKey)));
     }
 
     @Override
-    public String createKey(String walletName, WalletKeyType keyType){
-        return EosApiServiceGenerator.executeSync(eosWalletApiService.createKey(Arrays.asList(walletName, keyType.name())));
+    public String createKey(String walletName, WalletKeyType keyType) {
+        return EosApiServiceGenerator
+                .executeSync(eosWalletApiService.createKey(Arrays.asList(walletName, keyType.name())));
     }
 
     @Override
-    public Actions getActions(String accountName, Integer pos, Integer offset){
+    public Actions getActions(String accountName, Long pos, Long offset) {
         LinkedHashMap<String, Object> requestParameters = new LinkedHashMap<>(3);
 
         requestParameters.put("account_name", accountName);
@@ -219,7 +227,7 @@ public class EosApiRestClientImpl implements EosApiRestClient {
     }
 
     @Override
-    public Transaction getTransaction(String id){
+    public Transaction getTransaction(String id) {
         LinkedHashMap<String, String> requestParameters = new LinkedHashMap<>(1);
 
         requestParameters.put("id", id);
@@ -228,7 +236,7 @@ public class EosApiRestClientImpl implements EosApiRestClient {
     }
 
     @Override
-    public KeyAccounts getKeyAccounts(String publicKey){
+    public KeyAccounts getKeyAccounts(String publicKey) {
         LinkedHashMap<String, String> requestParameters = new LinkedHashMap<>(1);
 
         requestParameters.put("public_key", publicKey);
@@ -237,7 +245,7 @@ public class EosApiRestClientImpl implements EosApiRestClient {
     }
 
     @Override
-    public ControlledAccounts getControlledAccounts(String controllingAccountName){
+    public ControlledAccounts getControlledAccounts(String controllingAccountName) {
         LinkedHashMap<String, String> requestParameters = new LinkedHashMap<>(1);
 
         requestParameters.put("controlling_account", controllingAccountName);
